Convert useManageCloudBackups to Typescript and reorganize
